Implemented Loading data from API into RecipeName activity; Fixed ExoPlayer Bug
RecipeName activity loads a RecipeNameMasterListFragment. This is a Static Master List. This will be used to implement the Master/Detail Flow in android. Used Retrofit and helper classes and POJOs

Fixed the bug by checking if the savedInstanceState is null before creating a new Fragment to load the FragmentRecipePlayerView fragment into recipeStepDescriptionDetail activity using fragment manager