Bug on TwoPanelUI(To be fixed); Also Implemented populating the MainActivity class/Home page
Implemented populating the MainActivity as well as Sending the an Intent to the RecipeName class. I implemented a helper function that will return the Extras in the Intent which is a BakingFood list to the MasterListFragment class

FYI: the previous commit 'TwoPanel UI for tablet mode' has a few bugs which i fixed in this commit. But i have a new bug now where the video in the two panel UI wil not load when you select the step that has a video, but it does play when the tablet is in landscape mode. However, it does load the introduction video.