Fixed: Bug on TwoPanelUI; Also Implemented two UI polishing
The first UI polishing, was having the TwoPanelUI share equal width. The second UI polishing was displaying LinearLayout for the recyclerview of the MainActivity in mobile device, and having GridLayout for tablet mode

I fiex the bug by Instead of 'adding' a new Fragment, i had to 'replace' in the onRecipeStepDescriptionSelected method of RecipeName class.