Added a New App Widget
Created the App widget by following the IDE flow, as well as customizing the widget to open a pending intent to open the main Activity. The Ideal situation will be to open the recipeName showing ingredient and steps. This will be implemented. As well as also customizing the widget to refresh every 30 mins(1800000 mili seconds)