Testing RecyclerViews with Espresso for MainActivity
The test we are running is clicking on a RecyclerView baking food name and checking that it opens up the corresponding baking food ingredients and recipe steps. In particular, we selected the Yellow Cake in the list

We know it opens up the correct recipeName when the action bar title/text matches the name we specified. Test passed!